Rename comparison library
We don't want to give the impression that we've implemented the motoko
base library in azle.
