+++ examples/motoko_examples/quicksort/src/comparison.ts
