+++ examples/motoko_examples/quicksort/src/main.ts
@@ -2,8 +2,8 @@ import {
     Query,
     int
 } from 'azle';
-import { Int } from './motoko_base'
-import * as Quicksort from './quicksort'
+import { Int } from './comparison';
+import * as Quicksort from './quicksort';
 
 export function sort(xs: int[]): Query<int[]> {
     return Quicksort.sortBy(xs, Int.compare);
